fix: bring back kotlin_editor_export_plugin.cpp
